Compiling with gcc

+++ cg/src/graphics/GLRenderWindow2.cpp
@@ -28,7 +28,7 @@
 // Source file for OpenGL 2D render window.
 //
 // Author: Paulo Pagliosa
-// Last revision: 28/06/2023
+// Last revision: 30/07/2023
 
 #include "graphics/GLRenderWindow2.h"
 
@@ -142,7 +142,7 @@ GLRenderWindow2::mouseButtonInputEvent(int button, int actions, int)
 bool
 GLRenderWindow2::mouseMoveEvent(double xPos, double yPos)
 {
-  if (!_dragFlags)
+  if (!uint32_t(_dragFlags))
     return false;
   _mouseX = (int)xPos;
   _mouseY = (int)yPos;
